+++ torchbearer/metrics/decorators.py
@@ -4,12 +4,11 @@ here can be reproduced by manually interacting with the classes if necessary. Br
 used to construct a :class:`.MetricFactory` which will build a :class:`.MetricTree` that handles data flow between
 instances of :class:`.Mean`, :class:`.RunningMean`, :class:`.Std` etc.
 """
-
 import inspect
 
 from torchbearer import metrics
 
-from torchbearer.metrics import MetricFactory, EpochLambda, BatchLambda, ToDict, Mean, MetricTree, Std, RunningMean
+from torchbearer.metrics import EpochLambda, BatchLambda, ToDict, Mean, MetricTree, Std, RunningMean
 
 
 def default_for_key(key):
@@ -28,44 +27,35 @@ def default_for_key(key):
 
     """
     def decorator(arg):
-        if inspect.isclass(arg):
-            metric = arg()
-        else:
-            metric = arg
-        metrics.DEFAULT_METRICS[key] = metric
+        metrics.add_default(key, arg)
         return arg
     return decorator
 
 
 def lambda_metric(name, on_epoch=False):
     """The :func:`lambda_metric` decorator is used to convert a lambda function `y_pred, y_true` into a :class:`.Metric`
-    instance. In fact it return a :class:`.MetricFactory` which is used to build a :class:`.Metric`. This can make
-    things complicated as in the following example: ::
+    instance. This can be used as in the following example: ::
 
         @metrics.lambda_metric('my_metric')
         def my_metric(y_pred, y_true):
             ... # Calculate some metric
 
-        model = Model(metrics=[my_metric()]) # Note we have to call `my_metric` in order to instantiate the class
+        model = Model(metrics=[my_metric])
 
     :param name: The name of the metric (e.g. 'loss')
     :param on_epoch: If True the metric will be an instance of :class:`.EpochLambda` instead of :class:`.BatchLambda`
-    :return: A decorator which replaces a function with a :class:`.MetricFactory`
+    :return: A decorator which replaces a function with a :class:`.Metric`
     """
     def decorator(metric_function):
-        class LambdaFactory(MetricFactory):
-            def build(self):
-                if on_epoch:
-                    return EpochLambda(name, metric_function)
-                else:
-                    return BatchLambda(name, metric_function)
-
-        return LambdaFactory
+        if on_epoch:
+            return EpochLambda(name, metric_function)
+        else:
+            return BatchLambda(name, metric_function)
     return decorator
 
 
 def to_dict(clazz):
-    """The :func:`to_dict` decorator is used to wrap either a :class:`.Metric` or :class:`.MetricFactory` instance with
+    """The :func:`to_dict` decorator is used to wrap either a :class:`.Metric` class or a :class:`.Metric` instance with
     a :class:`.ToDict` instance. The result is that future output will be wrapped in a `dict[name, value]`.
 
     Example: ::
@@ -76,7 +66,7 @@ def to_dict(clazz):
         ... def my_metric(y_pred, y_true):
         ...     return y_pred + y_true
         ...
-        >>> my_metric().build().process({'y_pred':4, 'y_true':5})
+        >>> my_metric.process({'y_pred':4, 'y_true':5})
         9
 
         >>> @metrics.to_dict
@@ -84,25 +74,44 @@ def to_dict(clazz):
         ... def my_metric(y_pred, y_true):
         ...     return y_pred + y_true
         ...
-        >>> my_metric().build().process({'y_pred':4, 'y_true':5})
+        >>> my_metric.process({'y_pred':4, 'y_true':5})
         {'my_metric': 9}
 
     :param clazz: The class to *decorate*
-    :return: A :class:`.MetricFactory` which can be instantiated and built to wrap the given class in a :class:`.ToDict`
+    :return: A :class:`.ToDict` instance or a :class:`.ToDict` wrapper of the given class
     """
-    class DictFactory(MetricFactory):
-        def __init__(self, *args, **kwargs):
-            self.inner = clazz(*args, **kwargs)
+    if inspect.isclass(clazz):
+        class Wrapper(ToDict):
+            def __init__(self, *args, **kwargs):
+                super(Wrapper, self).__init__(clazz(*args, **kwargs))
+        Wrapper.__name__ = clazz.__name__
+        Wrapper.__doc__ = clazz.__doc__
+        return Wrapper
+    else:
+        return ToDict(clazz)
 
-        def build(self):
-            if isinstance(self.inner, MetricFactory):
-                inner = self.inner.build()
-            else:
-                inner = self.inner
 
-            return ToDict(inner)
+def _wrap_and_add_to_tree(clazz, child_func):
+    if inspect.isclass(clazz):
+        class Wrapper(MetricTree):
+            def __init__(self, *args, **kwargs):
+                inner = clazz(*args, **kwargs)
+                if isinstance(inner, MetricTree):
+                    super(Wrapper, self).__init__(inner.root)
+                    self.children = inner.children
+                else:
+                    super(Wrapper, self).__init__(inner)
 
-    return DictFactory
+                self.add_child(child_func(self.root))
+        Wrapper.__name__ = clazz.__name__
+        Wrapper.__doc__ = clazz.__doc__
+        return Wrapper
+    else:
+        inner = clazz
+        if not isinstance(inner, MetricTree):
+            inner = MetricTree(inner)
+        inner.add_child(child_func(inner))
+        return inner
 
 
 def mean(clazz):
@@ -117,10 +126,9 @@ def mean(clazz):
 
         >>> @metrics.mean
         ... @metrics.lambda_metric('my_metric')
-        ... def my_metric(y_pred, y_true):
+        ... def metric(y_pred, y_true):
         ...     return y_pred + y_true
         ...
-        >>> metric = my_metric().build()
         >>> metric.reset({})
         >>> metric.process({'y_pred':torch.Tensor([2]), 'y_true':torch.Tensor([2])}) # 4
         {}
@@ -132,25 +140,9 @@ def mean(clazz):
         {'my_metric': 6.0}
 
     :param clazz: The class to *decorate*
-    :return: A :class:`.MetricFactory` which can be instantiated and built to append a :class:`.Mean` to the\
-    :class:`.MetricTree`
+    :return: A :class:`.MetricTree` with a :class:`.Mean` appended or a wrapper class that extends :class:`.MetricTree`
     """
-    class MeanFactory(MetricFactory):
-        def __init__(self, *args, **kwargs):
-            self.inner = clazz(*args, **kwargs)
-
-        def build(self):
-            if isinstance(self.inner, MetricFactory):
-                inner = self.inner.build()
-            else:
-                inner = self.inner
-
-            if not isinstance(inner, MetricTree):
-                inner = MetricTree(inner)
-            inner.add_child(ToDict(Mean(inner.name)))
-            return inner
-
-    return MeanFactory
+    return _wrap_and_add_to_tree(clazz, lambda metric: ToDict(Mean(metric.name)))
 
 
 def std(clazz):
@@ -166,10 +158,9 @@ def std(clazz):
 
         >>> @metrics.std
         ... @metrics.lambda_metric('my_metric')
-        ... def my_metric(y_pred, y_true):
+        ... def metric(y_pred, y_true):
         ...     return y_pred + y_true
         ...
-        >>> metric = my_metric().build()
         >>> metric.reset({})
         >>> metric.process({'y_pred':torch.Tensor([2]), 'y_true':torch.Tensor([2])}) # 4
         {}
@@ -181,33 +172,15 @@ def std(clazz):
         '1.6330'
 
     :param clazz: The class to *decorate*
-    :return: A :class:`.MetricFactory` which can be instantiated and built to append a :class:`.Mean` to the\
-    :class:`.MetricTree`
+    :return: A :class:`.MetricTree` with a :class:`.Std` appended or a wrapper class that extends :class:`.MetricTree`
     """
-    class StdFactory(MetricFactory):
-        def __init__(self, *args, **kwargs):
-            self.inner = clazz(*args, **kwargs)
-
-        def build(self):
-            if isinstance(self.inner, MetricFactory):
-                inner = self.inner.build()
-            else:
-                inner = self.inner
-
-            if not isinstance(inner, MetricTree):
-                inner = MetricTree(inner)
-            inner.add_child(ToDict(Std(inner.name + '_std')))
-            return inner
-
-    return StdFactory
+    return _wrap_and_add_to_tree(clazz, lambda metric: ToDict(Std(metric.name + '_std')))
 
 
 def running_mean(clazz=None, batch_size=50, step_size=10):
-    """The :func:`running_mean` decorator is used to add a :class:`.RunningMean` to the :class:`.MetricTree`. As with the
-    other decorators, a :class:`.MetricFactory` is created which will do this upon the call to
-    :meth:`.MetricFactory.build`. If the inner class is not / does not build a :class:`.MetricTree` then one will be
-    created. The :class:`.RunningMean` will be wrapped in a :class:`.ToDict` (with 'running\_' prepended to the name) for
-    simplicity.
+    """The :func:`running_mean` decorator is used to add a :class:`.RunningMean` to the :class:`.MetricTree`. If the
+    inner class is not a :class:`.MetricTree` then one will be created. The :class:`.RunningMean` will be wrapped in a
+    :class:`.ToDict` (with 'running\_' prepended to the name) for simplicity.
 
     .. note::
         The decorator function does not need to be called if not desired, both: `@running_mean` and `@running_mean()`
@@ -220,10 +193,9 @@ def running_mean(clazz=None, batch_size=50, step_size=10):
 
         >>> @metrics.running_mean(step_size=2) # Update every 2 steps
         ... @metrics.lambda_metric('my_metric')
-        ... def my_metric(y_pred, y_true):
+        ... def metric(y_pred, y_true):
         ...     return y_pred + y_true
         ...
-        >>> metric = my_metric().build()
         >>> metric.reset({})
         >>> metric.process({'y_pred':torch.Tensor([2]), 'y_true':torch.Tensor([2])}) # 4
         {'running_my_metric': 4.0}
@@ -235,25 +207,11 @@ def running_mean(clazz=None, batch_size=50, step_size=10):
     :param clazz: The class to *decorate*
     :param batch_size: See :class:`.RunningMean`
     :param step_size: See :class:`.RunningMean`
-    :return: decorator or :class:`.MetricFactory`
+    :return: decorator or :class:`.MetricTree` instance or wrapper
     """
-    class RunningMeanFactory(MetricFactory):
-        def __init__(self, *args, **kwargs):
-            self.inner = clazz(*args, **kwargs)
-
-        def build(self):
-            if isinstance(self.inner, MetricFactory):
-                inner = self.inner.build()
-            else:
-                inner = self.inner
-
-            if not isinstance(inner, MetricTree):
-                inner = MetricTree(inner)
-            inner.add_child(ToDict(RunningMean('running_' + inner.name, batch_size=batch_size, step_size=step_size)))
-            return inner
-
     if clazz is None:
         def decorator(clazz):
             return running_mean(clazz, batch_size=batch_size, step_size=step_size)
         return decorator
-    return RunningMeanFactory
+
+    return _wrap_and_add_to_tree(clazz, lambda metric: ToDict(RunningMean('running_' + metric.name, batch_size=batch_size, step_size=step_size)))